refactor: ResultBase를 ResultContent로 인라인
- ResultBase.tsx 제거
- ResultBase의 모든 내용을 ResultContent.tsx로 통합
- 불필요한 컴포넌트 분리 제거로 코드 구조 단순화
